Revert "Dialog Overhaul and Boss Added"
This reverts commit 216ff9eb09576315eca5c8b01d57497b2c3e3edc.
