--- Pharmacy GO/Assets/Game/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e832008c38607874faec73cb30d265dd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
