--- Pharmacy GO/Assets/Game/UI/ChoiceText.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5cef80a9c2fdef347b016b800268a18c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
