--- Pharmacy GO/Assets/Scripts/Dialog.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 809233cff31631742b14caf4ede87237
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
