--- Pharmacy GO/Assets/Scripts/Dialog/ChoiceBox.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8f9d6bddbcdadc54995884f8d2733a40
