--- Pharmacy GO/Assets/Scripts/Dialog/ChoiceText.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 52cb18c410351f6489dadf4e09cbdc4e
