+++ Pharmacy GO/Assets/Scripts/GameControl.cs
@@ -21,7 +21,7 @@ public class GameController : MonoBehaviour
             state = GameState.Dialogue;
         };
 
-        DialogManager.Instance.OnDialogFinished += () =>
+        DialogManager.Instance.OnCloseDialog += () =>
         {
             if(state == GameState.Dialogue)
                 state = GameState.FreeRoam;
