+++ Pharmacy GO/Assets/Scripts/GamePlay/CoinManager.cs
@@ -27,16 +27,12 @@ public class CoinManager : MonoBehaviour
     public void AddCoin(int amount)
     {
         coinCount += amount;
+        Debug.Log("Coins: " + coinCount);
     }
 
     public int GetCoinCount()
     {
         return coinCount;
-    }
-
-    public void RemoveCoin(int amount)
-    {
-        coinCount -= amount;
     }
 }
