+++ Pharmacy GO/Assets/Scripts/GamePlay/Dialog.cs
@@ -1,4 +1,3 @@
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
