--- Pharmacy GO/Assets/Scripts/Player/Boss.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 43f312675fea6324c954213bc3074b74
