--- Pharmacy GO/Assets/Scripts/Player/BossController.cs
@@ -1,49 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class BossController : MonoBehaviour, Interactable
-{
-
-    public static BossController i { get; private set; }
-
-    private void Awake()
-    {
-        i = this;
-    }
-
-    public void Interact()
-    {
-        StartCoroutine(StartQuiz());
-    }
-
-    public IEnumerator StartQuiz()
-    {
-
-        int selectedChoice = 0;
-        yield return DialogManager.Instance.ShowDialogText("Are you ready to take the quiz?",
-            waitForInput: false,
-            choices: new List<string>() { "Yes", "No" },
-            onChoiceSelected: choiceIndex => selectedChoice = choiceIndex);
-
-        if (selectedChoice == 0)
-        {
-            //Yes
-            if (CoinManager.Instance.GetCoinCount() >= 10)
-            {
-                CoinManager.Instance.RemoveCoin(10);
-                // TODO: Send a call to Boss.cs to initiate a boss battle
-                StartCoroutine(DialogManager.Instance.ShowDialogText("Delicious coins om nom nom. You win!"));
-            }
-            else
-            {
-                StartCoroutine(DialogManager.Instance.ShowDialogText("You don't have enough coins! It takes 10 to try."));
-            }
-        }
-        else
-        {
-            //No
-            yield break;
-        }
-    }
-}
