--- Pharmacy GO/Assets/Scripts/Player/BossController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 80f5d2fef106aad4585cb916b656eb5a
